requirements.txt
gradio
pandas
numpy
python-dotenv
langchain
langchain-community
langchain-chroma
sentence-transformers
tf-keras
spotipy